feat: rot do boss

+++ Sources/Application/Handlers/Item/ItemHandler.cs
@@ -2427,8 +2427,8 @@ namespace NRO_Server.Application.Handlers.Item
                             disciple.Status = 4;
                         }
                         character.CharacterHandler.SetUpInfo();
-                        character.CharacterHandler.PlusHp((int)character.HpFull);
-                        character.CharacterHandler.PlusMp((int)character.MpFull);
+                        character.CharacterHandler.PlusHp(character.HpFull);
+                        character.CharacterHandler.PlusMp(character.MpFull);
                         character.CharacterHandler.SendZoneMessage(Service.UpdateBody(character));
                         character.CharacterHandler.SendMessage(Service.PlayerLoadSpeed(character));
                         character.CharacterHandler.SendMessage(Service.MeLoadPoint(character));
@@ -2527,8 +2527,8 @@ namespace NRO_Server.Application.Handlers.Item
                             disciple.Status = 4;
                         }
                         character.CharacterHandler.SetUpInfo();
-                        character.CharacterHandler.PlusHp((int)character.HpFull);
-                        character.CharacterHandler.PlusMp((int)character.MpFull);
+                        character.CharacterHandler.PlusHp(character.HpFull);
+                        character.CharacterHandler.PlusMp(character.MpFull);
                         character.CharacterHandler.SendZoneMessage(Service.UpdateBody(character));
                         character.CharacterHandler.SendMessage(Service.PlayerLoadSpeed(character));
                         character.CharacterHandler.SendMessage(Service.MeLoadPoint(character));
